# Source Generator Performance Improvements