--- a/src/Controls/src/SourceGen/Properties/launchSettings.json
+++ b/src/Controls/src/SourceGen/Properties/launchSettings.json
@@ -1,0 +1,8 @@
+{
+  "profiles": {
+    "SourceGen - Maui.Controls.Sample": {
+      "commandName": "DebugRoslynComponent",
+      "targetProject": "..\\..\\samples\\Controls.Sample\\Maui.Controls.Sample.csproj"
+    }
+  }
+}