### Description of Change

<!-- Enter description of the fix in this section -->

Source Generator Performance Improvements:
- Reversed lookup order (Extension second)
- Introduced type cache
- Restored lookup order (Extension first again)

```
PERF PROGRESS
Updated numbers. Not quite as dramatic but still a good improvement. The original measurements had a TracePoint skewing the timing results quite heavily. The new numbers only use a StopWatch in code with the TracePoint disabled.

SourceGen - Maui.Controls.Sample (262 XAML files)

1) Unchanged:
- 15640 GetTypeByMetadata calls
~ 4s

2) Extensions lookup in XmlTypeXamlExtensions.GetTypeReference() second
- 6232 GetTypeByMetadata calls (~60% reduction)
~ 2.6s                         (~35% reduction)

3) With type cache and Extension lookup second
- 203 GetTypeByMetadata calls  (~99% reduction)
~ 1.5s                         (~62% reduction => ~2.6 times faster)

4) With just a type cache
- 523 GetTypeByMetadata calls  (~97% reduction)
~ 1.5s                         (~62% reduction => ~2.6 times faster)
```

### Issues Fixed

<!-- Please make sure that there is a bug logged for the issue being fixed. The bug should describe the problem and how to reproduce it. -->

Fixes #19944 GetTypeNameFromCustomNamespace does not cache duplicate results

<!--
Are you targeting main? All PRs should target the main branch unless otherwise noted.
-->