@mgoertz-msft 2024-01-26
@rmarinho Could you please review again? I had to switch the Extension lookup order back because it preferred System.Array over ArrayExtension for x:Array, and it's probably safer to keep this way in case there are other scenarios like this that we may not be aware about. The cache brought the real improvement so this lookup switch wasn't needed anyway.