@mgoertz-msft 2024-01-26
> I wonder if we could add a Benchmark around this change?

How do you do that? Do you have an example?