@jonathanpeppers 2024-01-26
There is a BenchmarkDotNet project, but it's not setup to test things like source generators or MSBuild tasks:

https://github.com/dotnet/maui/tree/main/src/Core/tests/Benchmarks/Benchmarks

In theory, we could make a `Benchmarks.Build` or `Benchmarks.Tooling` project for things like this.